Convert StringToObjectConverter and implementations to top-level types
