+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/converter/DefaultArgumentConverter.java
@@ -12,46 +12,23 @@ package org.junit.jupiter.params.converter;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.unmodifiableList;
-import static java.util.Collections.unmodifiableMap;
 import static org.apiguardian.api.API.Status.INTERNAL;
 import static org.junit.platform.commons.util.ReflectionUtils.getWrapperType;
 
 import java.io.File;
 import java.math.BigDecimal;
 import java.math.BigInteger;
-import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
-import java.nio.charset.Charset;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.time.Duration;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.MonthDay;
-import java.time.OffsetDateTime;
-import java.time.OffsetTime;
-import java.time.Period;
-import java.time.Year;
-import java.time.YearMonth;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
 import java.util.Currency;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.function.Function;
 
 import org.apiguardian.api.API;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.platform.commons.util.ClassLoaderUtils;
-import org.junit.platform.commons.util.Preconditions;
 import org.junit.platform.commons.util.ReflectionUtils;
 
 /**
@@ -141,199 +118,6 @@ public class DefaultArgumentConverter implements ArgumentConverter {
 	private static Class<?> toWrapperType(Class<?> targetType) {
 		Class<?> wrapperType = getWrapperType(targetType);
 		return wrapperType != null ? wrapperType : targetType;
-	}
-
-	interface StringToObjectConverter {
-
-		/**
-		 * Determine if this converter can convert from a {@link String} to the
-		 * supplied target type (which is guaranteed to be a wrapper type for
-		 * primitives &mdash; for example, {@link Integer} instead of {@code int}).
-		 */
-		boolean canConvert(Class<?> targetType);
-
-		Object convert(String source, Class<?> targetType) throws Exception;
-
-		default Object convert(String source, Class<?> targetType, ClassLoader classLoader) throws Exception {
-			return convert(source, targetType);
-		}
-
-	}
-
-	private static class StringToBooleanConverter implements StringToObjectConverter {
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return targetType == Boolean.class;
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) {
-			boolean isTrue = "true".equalsIgnoreCase(source);
-			Preconditions.condition(isTrue || "false".equalsIgnoreCase(source),
-				() -> "String must be 'true' or 'false' (ignoring case): " + source);
-			return isTrue;
-		}
-	}
-
-	private static class StringToCharacterConverter implements StringToObjectConverter {
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return targetType == Character.class;
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) {
-			Preconditions.condition(source.length() == 1, () -> "String must have length of 1: " + source);
-			return source.charAt(0);
-		}
-	}
-
-	private static class StringToNumberConverter implements StringToObjectConverter {
-
-		private static final Map<Class<?>, Function<String, ?>> CONVERTERS;
-		static {
-			Map<Class<?>, Function<String, ?>> converters = new HashMap<>();
-			converters.put(Byte.class, Byte::decode);
-			converters.put(Short.class, Short::decode);
-			converters.put(Integer.class, Integer::decode);
-			converters.put(Long.class, Long::decode);
-			converters.put(Float.class, Float::valueOf);
-			converters.put(Double.class, Double::valueOf);
-			CONVERTERS = unmodifiableMap(converters);
-		}
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return CONVERTERS.containsKey(targetType);
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) {
-			return CONVERTERS.get(targetType).apply(source.replace("_", ""));
-		}
-	}
-
-	private static class StringToClassConverter implements StringToObjectConverter {
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return targetType == Class.class;
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) throws Exception {
-			throw new UnsupportedOperationException();
-		}
-
-		@Override
-		public Object convert(String className, Class<?> targetType, ClassLoader classLoader) throws Exception {
-			// @formatter:off
-			return ReflectionUtils.tryToLoadClass(className, classLoader)
-					.getOrThrow(cause -> new ArgumentConversionException(
-							"Failed to convert String \"" + className + "\" to type java.lang.Class", cause));
-			// @formatter:on
-		}
-
-	}
-
-	private static class StringToEnumConverter implements StringToObjectConverter {
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return targetType.isEnum();
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) throws Exception {
-			return valueOf(targetType, source);
-		}
-
-		@SuppressWarnings({ "unchecked", "rawtypes" })
-		private Object valueOf(Class targetType, String source) {
-			return Enum.valueOf(targetType, source);
-		}
-	}
-
-	private static class StringToJavaTimeConverter implements StringToObjectConverter {
-
-		private static final Map<Class<?>, Function<String, ?>> CONVERTERS;
-		static {
-			Map<Class<?>, Function<String, ?>> converters = new HashMap<>();
-			converters.put(Duration.class, Duration::parse);
-			converters.put(Instant.class, Instant::parse);
-			converters.put(LocalDate.class, LocalDate::parse);
-			converters.put(LocalDateTime.class, LocalDateTime::parse);
-			converters.put(LocalTime.class, LocalTime::parse);
-			converters.put(MonthDay.class, MonthDay::parse);
-			converters.put(OffsetDateTime.class, OffsetDateTime::parse);
-			converters.put(OffsetTime.class, OffsetTime::parse);
-			converters.put(Period.class, Period::parse);
-			converters.put(Year.class, Year::parse);
-			converters.put(YearMonth.class, YearMonth::parse);
-			converters.put(ZonedDateTime.class, ZonedDateTime::parse);
-			converters.put(ZoneId.class, ZoneId::of);
-			converters.put(ZoneOffset.class, ZoneOffset::of);
-			CONVERTERS = unmodifiableMap(converters);
-		}
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return CONVERTERS.containsKey(targetType);
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) throws Exception {
-			return CONVERTERS.get(targetType).apply(source);
-		}
-	}
-
-	private static class StringToCommonJavaTypesConverter implements StringToObjectConverter {
-
-		private static final Map<Class<?>, Function<String, ?>> CONVERTERS;
-
-		static {
-			Map<Class<?>, Function<String, ?>> converters = new HashMap<>();
-
-			// java.io and java.nio
-			converters.put(File.class, File::new);
-			converters.put(Charset.class, Charset::forName);
-			converters.put(Path.class, Paths::get);
-			// java.net
-			converters.put(URI.class, URI::create);
-			converters.put(URL.class, StringToCommonJavaTypesConverter::toURL);
-			// java.math
-			converters.put(BigDecimal.class, BigDecimal::new);
-			converters.put(BigInteger.class, BigInteger::new);
-			// java.util
-			converters.put(Currency.class, Currency::getInstance);
-			converters.put(Locale.class, Locale::new);
-			converters.put(UUID.class, UUID::fromString);
-
-			CONVERTERS = unmodifiableMap(converters);
-		}
-
-		@Override
-		public boolean canConvert(Class<?> targetType) {
-			return CONVERTERS.containsKey(targetType);
-		}
-
-		@Override
-		public Object convert(String source, Class<?> targetType) throws Exception {
-			return CONVERTERS.get(targetType).apply(source);
-		}
-
-		private static URL toURL(String url) {
-			try {
-				return URI.create(url).toURL();
-			}
-			catch (MalformedURLException ex) {
-				throw new ArgumentConversionException("Failed to convert String \"" + url + "\" to type java.net.URL",
-					ex);
-			}
-		}
-
 	}
 
 }
